implemented the Loading for the epdb v1 implementation
