--- EasePass/Core/Database/epdb/v1/DatabaseLoader.cs
@@ -1,17 +1,0 @@
-﻿namespace EasePass.Core.Database.epdb.v1
-{
-    internal class DatabaseLoader : ADatabaseLoader, IDatabaseLoader
-    {
-        public static double Version => 1.0;
-
-        public static string Decrypt()
-        {
-            throw new System.NotImplementedException();
-        }
-
-        public static byte[] Encrypt()
-        {
-            throw new System.NotImplementedException();
-        }
-    }
-}
